Updated for NPE
When the welcome page is loaded for the first time, the
allLocationStatsMap is set with all the entries from the API. But while
rendering onto the UI, the values corresponding to "Total" are removed
as they're redundant. But this was also removing the value permanently
from the allLocationStatsMap object. Fixed this issue.